stworzone formularze dodawania uczniow i nauczycieli

+++ korepetycje2/src/main/java/org/example/lesson/LessonRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LessonRepository extends JpaRepository<Lesson,Integer> {
+public interface LessonRepository extends JpaRepository<Lesson,Long> {
 }
